Add feature: Switch hides banner image in list
Adds the functionality for a user to include an extra parameter in their post header to hide the banner image in a post list.

Currently, if an image is defined in the header, like `image = "example.png"`, then the banner image will appear in both the post and the post list. It is not possible (to my knowledge) to remove the image from the post list, but keep it in the actual post. This change would add this functionality.

The naming I used for this new switch probably isn't great. It's a little long. However, I couldn't think of anything much better that would still be descriptive and understandable. Please change if you think of something better.

Cheers!