Support external header images.
Previous implimentation hard-coded an "img/" prefix to the
header image URL. This means posts cannot use images from
an external source.

This first parses the URL and checks if it has a http/https
prefix. If it does, we know it's an absolute URL, so we don't
add any prefixes.

Otherwise it's a relative URL, and we append the "img/" prefix
as it did before.