Bugfix: missing see all posts link
After the "Offset and Limit" section, `$query` would have length of `$items_count`, which makes `$count` always equal to `$items_count`. The later "if" condition for adding the "see all posts" link will never be met.

The fix is to record the length of `$query` into `$count` before any offset and limit.