Add option to use self hosted commento instance
I wanted to use a self hosted instance of commento. I could have simply
changed the url in the html file but i thought this could be useful for
others too. If you are using the commento cloud service, you do not have to
change anything, but if you want to use a self hosted instance, you now
just have to supply the base url instead of changing theme files.