feat: add options to customise length of listing summaries
When `summary` is defined, the full summary text is now shown in listings.

Add `summaryLength = 30` (length in words) to `config.toml` to customise automatic summaries from using `<!--more-->` tag or full content (when `summary` and `abstract` are not defined).

Add `abstract_length = 135` (length in characters) to `params.toml` to customise truncation when `abstract` is used as the summary.

Close #1374