fix: social sharing image for author pages
The social sharing logic was attempting to use the legacy system of defining the avatar image name in `site.Params.avatar`, however since the author system was added to Academic, `site.Params.avatar` is now a group of params and the avatar itself is stored in the author's page bundle.

Hence, use the avatar in the author's page bundle (if it exists) to represent an author page.

This bug could affect the image used to represent a page in social sharing and RSS.