feat(widget): show email (if entered) in About widget
Previously email only shown if exists and the email form is disabled.

However, both Formspree and Netlify no longer use email address as a parameter, and received request to show email (if entered), in addition to email form, so visitors have the choice to use their usual email client to send a message.

Therefore, show the email address if it was entered.