starters(academic): update default cms branch (#2705)
The Academic repo branch was renamed from master to main so when netlify creates a new repo based on this template, the default branch is named 'main' instead of 'master'.

Leaving the default cms branch param as 'master' will result in a credentials error whenever a user tries to login using the /admin page. Renaming this to 'main' fixes the error.