Add missing Reddit 'reddit-alien' icon
The default 'page_sharer.yaml' file mentions the icon 'reddit-alien' in the Reddit section (which is commented out by default). However, the above mentioned icon is not found in the 'brands' collection (or in any other collections, as far as I could see). This commit adds (back) the 'reddit-alien' icon.